Moved native/request.Protocol (copied from openrtb.Protocol) to native.Protocol (for consistency)
